cognix-flask-server/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
groq==0.11.0
pandas==2.2.3
numpy==2.0.2
scikit-learn==1.5.2
scipy==1.14.1
joblib==1.4.2
requests
xgboost==2.0.3
lightgbm==4.3.0

matplotlib==3.9.3
seaborn==0.13.2

gunicorn==21.2.0
reportlab==4.0.7
python-dotenv==1.0.0
cloudinary
httpx<0.28.0
